+++ sdk/netapp/Microsoft.Azure.Management.NetApp/src/generate.ps1
@@ -1,1 +1,0 @@
-Start-AutoRestCodeGeneration -ResourceProvider "netapp/resource-manager" -AutoRestVersion "v2" -SpecsRepoBranch "main" -AutoRestCodeGenerationFlags "--use:@microsoft.azure/autorest.csharp@2.3.90"