+++ sdk/sqlvirtualmachine/Microsoft.Azure.Management.SqlVirtualMachine/src/generate.ps1
@@ -1,1 +1,0 @@
-Start-AutoRestCodeGeneration -ResourceProvider "sqlvirtualmachine/resource-manager" -AutoRestVersion "v2"