+++ sdk/storage/Azure.Storage.Blobs/tests/BlobBaseClientTests.cs
@@ -5111,6 +5111,83 @@ namespace Azure.Storage.Blobs.Test
         }
 
         [RecordedTest]
+        public async Task SetMetadataAsync_Sort()
+        {
+            await using DisposingContainer test = await GetTestContainerAsync();
+            // Arrange
+            BlobBaseClient blob = await GetNewBlobClient(test.Container);
+            IDictionary<string, string> metadata = new Dictionary<string, string>() {
+                { "a0", "a" },
+                { "a1", "a" },
+                { "a2", "a" },
+                { "a3", "a" },
+                { "a4", "a" },
+                { "a5", "a" },
+                { "a6", "a" },
+                { "a7", "a" },
+                { "a8", "a" },
+                { "a9", "a" },
+                { "_", "a" },
+                { "_a", "a" },
+                { "a_", "a" },
+                { "__", "a" },
+                { "_a_", "a" },
+                { "b", "a" },
+                { "c", "a" },
+                { "y", "a" },
+                { "z", "z_" },
+                { "_z", "a" },
+                { "_F", "a" },
+                { "F", "a" },
+                { "F_", "a" },
+                { "_F_", "a" },
+                { "__F", "a" },
+                { "__a", "a" },
+                { "a__", "a" }
+             };
+
+            // Act
+            Response<BlobInfo> response = await blob.SetMetadataAsync(metadata);
+
+            // Assert
+
+            // Ensure that we grab the whole ETag value from the service without removing the quotes
+            Assert.AreEqual(response.Value.ETag.ToString(), $"\"{response.GetRawResponse().Headers.ETag}\"");
+
+            // Ensure the value has been correctly set by doing a GetProperties call
+            Response<BlobProperties> getPropertiesResponse = await blob.GetPropertiesAsync();
+            AssertDictionaryEquality(metadata, getPropertiesResponse.Value.Metadata);
+        }
+
+        [RecordedTest]
+        public async Task SetMetadataAsync_Sort_InvalidMetadata()
+        {
+            await using DisposingContainer test = await GetTestContainerAsync();
+            // Arrange
+            BlobBaseClient blob = await GetNewBlobClient(test.Container);
+            IDictionary<string, string> metadata = new Dictionary<string, string>() {
+                { "test", "val" },
+                { "test-", "val" },
+                { "test--", "val" },
+                { "test-_", "val" },
+                { "test_-", "val" },
+                { "test__", "val" },
+                { "test-a", "val" },
+                { "test-_A", "val" },
+                { "test_a", "val" },
+                { "test_Z", "val" },
+                { "test_a_", "val" },
+                { "test_a-", "val" },
+                { "test_a-_", "val" },
+             };
+
+            // Act
+            await TestHelper.AssertExpectedExceptionAsync<RequestFailedException>(
+                blob.SetMetadataAsync(metadata),
+                e => Assert.AreEqual(BlobErrorCode.InvalidMetadata.ToString(), e.ErrorCode));
+        }
+
+        [RecordedTest]
         public async Task CreateSnapshotAsync()
         {
             await using DisposingContainer test = await GetTestContainerAsync();
@@ -7132,8 +7209,10 @@ namespace Azure.Storage.Blobs.Test
                 constants.Sas.SharedKeyCredential,
                 GetOptions()));
 
+            string stringToSign = null;
+
             // Act
-            Uri sasUri = blobClient.GenerateSasUri(permissions, expiresOn);
+            Uri sasUri = blobClient.GenerateSasUri(permissions, expiresOn, out stringToSign);
 
             // Assert
             BlobSasBuilder sasBuilder = new BlobSasBuilder(permissions, expiresOn)
@@ -7148,6 +7227,7 @@ namespace Azure.Storage.Blobs.Test
                 Sas = sasBuilder.ToSasQueryParameters(constants.Sas.SharedKeyCredential)
             };
             Assert.AreEqual(expectedUri.ToUri(), sasUri);
+            Assert.IsNotNull(stringToSign);
         }
 
         [RecordedTest]
@@ -7177,8 +7257,10 @@ namespace Azure.Storage.Blobs.Test
                 StartsOn = startsOn
             };
 
+            string stringToSign = null;
+
             // Act
-            Uri sasUri = blobClient.GenerateSasUri(sasBuilder);
+            Uri sasUri = blobClient.GenerateSasUri(sasBuilder, out stringToSign);
 
             // Assert
             BlobSasBuilder sasBuilder2 = new BlobSasBuilder(permissions, expiresOn)
@@ -7194,6 +7276,7 @@ namespace Azure.Storage.Blobs.Test
                 Sas = sasBuilder2.ToSasQueryParameters(constants.Sas.SharedKeyCredential)
             };
             Assert.AreEqual(expectedUri.ToUri(), sasUri);
+            Assert.IsNotNull(stringToSign);
         }
 
         [RecordedTest]