+++ sdk/storage/Azure.Storage.Common/CHANGELOG.md
@@ -3,12 +3,18 @@
 ## 12.21.0-beta.1 (Unreleased)
 
 ### Features Added
+- Added ability to retrieve SAS string to sign for debugging purposes.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 12.20.1 (2024-07-25)
+
+### Bugs Fixed
+- Fixed \[BUG\] Azure Blob Storage Client SDK No Longer Supports Globalization Invariant Mode for Account Key Authentication #45052
 
 ## 12.20.0 (2024-07-16)